Extract shared SSO callback handler for Okta and Entra
